+++ src/components/App/App.js
@@ -2,7 +2,7 @@ import React from "react";
 
 import ToastPlayground from "../ToastPlayground";
 import Footer from "../Footer";
-import ToastProvider from "../ToastProvider/ToastProvider";
+import ToastProvider from "../ToastProvider";
 
 function App() {
   return (
